“Stand out in a crowd! Understand what Data Science is and how it can boost your profile”
This workshop will give participants a sense of what Data Science is and what the different parts of the data cycle are. This workshop will provide practical examples from different domains that explain how usage of some or all parts of the data science cycle may benefit you in your academic work and professional careers.