“You will spend nearly 70% of your time doing this! Organize & pre-process your Data”
An important part of data science is the work needed to prepare data for use. This workshop will focus on the importance and the techniques used for cleaning, integration, pre-processing, and transformation of raw data. We will use examples from cross-country survey data to see how unclean data can pose serious problems, and then use different methods to treat and transform this data into a meaningful format. No previous experience is needed.