“The Power of Visual Storytelling – Learning Tableau Public”
This workshop will explore the capabilities of Tableau Public – a user-friendly, freely available tool that can be used to create impactful visualizations with a click of a button using any amount of data. No previous experience is necessary.
Note: Instructions for downloading Tableau will be sent prior to the workshop.