“Part I - Make your computer work for you! Learn how to use Python to program your tasks”
This workshop will introduce the idea of using programming languages to delegate tasks to your machine and demonstrate the benefits and ease of use that Python offers. We will cover Python basics, an introduction to the rich library of data science resources in Python, and an introduction to Jupyter Notebooks as a powerful tool for coding and documenting your projects. No previous experience is needed.