“The Chronicles of Data Science: An end-to-end project”
We will do it all! This workshop will take a particular example problem and solve it using different data science techniques. We will pre-process, clean, store, visualize and create models with our data for predictions! We will use the popular and freely available Python3 & Jupyter Notebooks technologies in this workshop, but no prior experience is necessary.